move libwallet error into separate file
