+++ wallet/src/client.rs
@@ -22,6 +22,7 @@ use serde_json;
 use tokio_core::reactor;
 
 use libwallet::types::*;
+use libwallet::error::{Error, ErrorKind};
 use std::io;
 use util::LOGGER;
