+++ wallet/src/libwallet/keys.rs
@@ -14,7 +14,8 @@
 
 //! Wallet key management functions
 use keychain::Identifier;
-use libwallet::types::{Error, WalletBackend};
+use libwallet::types::WalletBackend;
+use libwallet::error::Error;
 
 /// Get our next available key
 pub fn new_output_key<T>(wallet: &mut T) -> Result<(Identifier, u32), Error>
