+++ wallet/src/libwallet/mod.rs
@@ -29,4 +29,5 @@ pub mod keys;
 pub mod selection;
 pub mod sigcontext;
 pub mod types;
+pub mod error;
 pub mod updater;
