+++ wallet/src/sender.rs
@@ -19,7 +19,8 @@ use core::ser;
 use failure::ResultExt;
 use keychain::{Identifier, Keychain};
 use libtx::{build, tx_fee};
-use libwallet::types::{Error, ErrorKind, WalletBackend};
+use libwallet::error::{Error, ErrorKind};
+use libwallet::types::WalletBackend;
 use libwallet::{selection, updater};
 use receiver::TxWrapper;
 use util;
