+++ wallet/tests/common/mod.rs
@@ -29,6 +29,7 @@ use core::core::{Output, OutputFeatures, OutputIdentifier, Transaction, TxKernel
 use core::{consensus, global, pow};
 use wallet::file_wallet::*;
 use wallet::libwallet::types::*;
+use wallet::libwallet::error::*;
 use wallet::libwallet::updater;
 
 use util::secp::pedersen;
